requirements.txt
kokoro
soundfile
torch
python-vlc
tqdm
